acdb: ats: Drop libdiag dependency and use dlsym instead
libdiag is a proprietary library and we do not have its source code,
but the library itself is supplied with the ODM. This means we can
dynamically link it during runtime. Add the load_diag_lib() function
that loads the dynamic library and gets the addresses of the necessary
functions.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>